Termux Client/requirements.txt
marshmallow==3.21.1
packaging==24.0
paho-mqtt==2.0.0
termux-api==1.2.5.2
